added post renewal webhook + copy certs in another folder + update ca-certicates store from time to time
Webhook is to tell the service to reload the new certificate. Could be a web server or portainer for example

I also copy the certs in another folder with a constant cert.crt and cert.key name for easier integration in other configs

The service also update its ca-certificates store from time to time. Ideally this container should be able to run for years without update

This should have been 3 commits but I did not plan to upload that back at first